requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.1
pycodestyle==2.9.1
pykerberos==1.2.4
pymongo==4.2.0
python-snappy==0.6.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
